declutter listing filter logic
